separated equipment into two files
